+++ test/e2e/authorization_server.spec.ts
@@ -14,7 +14,7 @@ import {
   AuthorizationServer,
   base64encode,
   DateInterval,
-  IAuthCodePayload,
+  PayloadAuthCode,
   JwtService,
   OAuthClient,
   OAuthRequest,
@@ -187,7 +187,7 @@ describe("authorization_server", () => {
 
     const response = await authorizationServer.completeAuthorizationRequest(authorizationRequest);
     const authorizeResponseQuery = new URLSearchParams(response.headers.location.split("?")[1]);
-    const decodedCode: IAuthCodePayload = <IAuthCodePayload>decode(String(authorizeResponseQuery.get("code")));
+    const decodedCode: PayloadAuthCode = <PayloadAuthCode>decode(String(authorizeResponseQuery.get("code")));
 
     expect(decodedCode.client_id).toBe(client.id);
     expect(decodedCode.redirect_uri).toBe("http://localhost");
@@ -256,7 +256,7 @@ describe("authorization_server", () => {
 
       // assert
       const authorizeResponseQuery = new URLSearchParams(response.headers.location.split("?")[1]);
-      const decodedCode: IAuthCodePayload = <IAuthCodePayload>decode(String(authorizeResponseQuery.get("code")));
+      const decodedCode: PayloadAuthCode = <PayloadAuthCode>decode(String(authorizeResponseQuery.get("code")));
       expect(decodedCode.client_id).toBe(client.id);
       expect(decodedCode.redirect_uri).toBe("http://localhost");
       expect(decodedCode.code_challenge).toBeUndefined();